Added functionality of the Diagnosis.java class. test results now are saved in the excel files + Added gender field to Patient.java. changed hashtable type to String Double + Added right to left slide transition animations in SlideTransitions.java

+++ src/main/Application/main/InsertData.java
@@ -45,7 +45,6 @@ public class InsertData implements Initializable {
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
-
     }
 
     public void homeButton_Click(ActionEvent e) throws IOException {
@@ -61,6 +60,7 @@ public class InsertData implements Initializable {
     }
 
     public void saveButton_Click(ActionEvent e) throws IOException {
+
         if (checkField(whiteBloodCellsField) &&
                 checkField(neutrophilField) &&
                 checkField(lymphocytesField) &&
@@ -72,26 +72,24 @@ public class InsertData implements Initializable {
                 checkField(lipoproteinField) &&
                 checkField(phophataseField) &&
                 checkField(redBloodCellsField)) {
-            Hashtable<String, Integer> bloodTest = new Hashtable<>();
-            bloodTest.put("WBC", Integer.valueOf(whiteBloodCellsField.getText()));
-            bloodTest.put("Neut", Integer.valueOf(neutrophilField.getText()));
-            bloodTest.put("Lymph", Integer.valueOf(lymphocytesField.getText()));
-            bloodTest.put("RBC", Integer.valueOf(redBloodCellsField.getText()));
-            bloodTest.put("HCT", Integer.valueOf(hematocritField.getText()));
-            bloodTest.put("Urea", Integer.valueOf(ureaField.getText()));
-            bloodTest.put("Hb", Integer.valueOf(hemoglobinField.getText()));
-            bloodTest.put("Crtn", Integer.valueOf(creatinineField.getText()));
-            bloodTest.put("Iron", Integer.valueOf(ironField.getText()));
-            bloodTest.put("HDL", Integer.valueOf(lipoproteinField.getText()));
-            bloodTest.put("AP", Integer.valueOf(phophataseField.getText()));
-
+            Hashtable<String, Double> bloodTest = new Hashtable<>();
+            bloodTest.put("WBC", Double.valueOf(whiteBloodCellsField.getText()));
+            bloodTest.put("Neut", Double.valueOf(neutrophilField.getText()));
+            bloodTest.put("Lymph", Double.valueOf(lymphocytesField.getText()));
+            bloodTest.put("RBC", Double.valueOf(redBloodCellsField.getText()));
+            bloodTest.put("HCT", Double.valueOf(hematocritField.getText()));
+            bloodTest.put("Urea", Double.valueOf(ureaField.getText()));
+            bloodTest.put("Hb", Double.valueOf(hemoglobinField.getText()));
+            bloodTest.put("Crtn", Double.valueOf(creatinineField.getText()));
+            bloodTest.put("Iron", Double.valueOf(ironField.getText()));
+            bloodTest.put("HDL", Double.valueOf(lipoproteinField.getText()));
+            bloodTest.put("AP", Double.valueOf(phophataseField.getText()));
             Node node = (Node) e.getSource();
             Stage stage = (Stage) node.getScene().getWindow();
             Patient p = (Patient) stage.getUserData();
             p.setBloodTest(bloodTest);
         }
         new SlideTransitions().leftToRightTransition(parentContainer, homeButton, anchorPane, "Diagnosis.fxml");
-
 
     }
 
@@ -114,22 +112,17 @@ public class InsertData implements Initializable {
     public void handleDrop(DragEvent dragEvent) throws FileNotFoundException {
         path = String.valueOf(dragEvent.getDragboard().getFiles());
         path = path.replace("[", "").replace("]", "");
-        //   xlsxfile= new File( dragEvent.getDragboard().getFiles().toString());
         drop_text.setText("Done");
     }
 
-
     public void dropSaveButton_Click(ActionEvent actionEvent) {
         if (path != null) {
-
             Node node = (Node) actionEvent.getSource();
             Stage stage = (Stage) node.getScene().getWindow();
             Patient p = (Patient) stage.getUserData();
             p.addBloodTest(path);
             System.out.println(p.getFirstName());
-
             System.out.println(p.getBloodTest().toString());
-
         } else drop_text.setText("Error");
     }
 }
